sistema-ortomedica/Sistema_Ortomedica_v4/requirements.txt
playwright==1.47.0
python-dotenv==1.0.1
mysql-connector-python==9.0.0
openpyxl==3.1.2
xlrd==1.2.0
fastapi
uvicorn
mysql-connector-python
python-jose
passlib[bcrypt]
pytz
